feat: add automatic text chunking with aggregate embeddings
Add automatic text chunking for documents exceeding embedding model token limits:

- Create chunker package with configurable chunking strategies
  - FixedOverlapChunker: splits text with configurable size and overlap
  - Uses tiktoken for accurate token counting (cl100k_base encoding)
  - Default: 512 token chunks with 50 token overlap, 8191 max tokens

- Implement batch embedding support
  - Add BatchEmbeddingProvider interface for efficient multi-text embedding
  - Implement EmbedBatch in OpenAI provider (up to 2048 texts per request)
  - Automatic fallback to individual embeddings if batch not supported

- Aggregate embedding approach
  - Chunk text when exceeding token limits
  - Embed all chunks using batch API for performance
  - Average chunk embeddings into single aggregate embedding
  - Store as single entry (no derived keys needed)

- Convert entire codebase from float32 to float64
  - Use OpenAI's native float64 format (eliminates conversions)
  - Update all similarity functions and backends
  - Better precision for similarity calculations

- Configuration options
  - Chunking enabled by default with sensible defaults
  - WithChunking() to customize chunk size/overlap/strategy
  - WithoutChunking() to disable if not needed
  - Zero-config works out of the box

- Add comprehensive tests
  - Chunker package tests (validation, chunking logic, token counting)
  - Cache integration tests (aggregate embeddings, enable/disable)
  - All existing tests pass

+++ options/options_test.go
@@ -13,11 +13,11 @@ type mockProvider struct {
 	shouldErr bool
 }
 
-func (m *mockProvider) EmbedText(text string) ([]float32, error) {
+func (m *mockProvider) EmbedText(text string) ([]float64, error) {
 	if m.shouldErr {
 		return nil, &testError{"mock error"}
 	}
-	return []float32{0.1, 0.2, 0.3}, nil
+	return []float64{0.1, 0.2, 0.3}, nil
 }
 
 func (m *mockProvider) Close() {}
@@ -169,7 +169,7 @@ func TestProviderOptions(t *testing.T) {
 func TestSimilarityOptions(t *testing.T) {
 	t.Run("CustomSimilarity", func(t *testing.T) {
 		cfg := NewConfig[string, string]()
-		customSim := func(a, b []float32) float32 { return 0.5 }
+		customSim := func(a, b []float64) float64 { return 0.5 }
 
 		err := cfg.Apply(WithSimilarityComparator[string, string](customSim))
 		if err != nil {
@@ -177,7 +177,7 @@ func TestSimilarityOptions(t *testing.T) {
 		}
 
 		// Test that the custom function works
-		result := cfg.Comparator([]float32{1, 0}, []float32{0, 1})
+		result := cfg.Comparator([]float64{1, 0}, []float64{0, 1})
 		if result != 0.5 {
 			t.Errorf("Expected 0.5, got %f", result)
 		}
@@ -246,7 +246,7 @@ func (m *mockBackend[K, V]) Flush(ctx context.Context) error {
 	return nil
 }
 
-func (m *mockBackend[K, V]) GetEmbedding(ctx context.Context, key K) ([]float32, bool, error) {
+func (m *mockBackend[K, V]) GetEmbedding(ctx context.Context, key K) ([]float64, bool, error) {
 	return nil, false, nil
 }
